Use a macro for kbd commands
Texinfo uses the @kbd macro to markup commands to be entered at the
keyboard. This is quite cumbersome to add via @@texinfo: commands inside
org. Luckily there are org mode MACROS which help to somewhat make this
more readable.

OTOH I guess we could just use ~M-x package-install-file~ but this
results in @code macros in texinfo which is not quite what we want.